Added Windows port.

+++ src/System/Common.h
@@ -1,12 +1,12 @@
 #ifndef COMMON_H
 #define COMMON_H
 
-#include "assert.h"
+#include <assert.h>
 #include <string.h>
-#include <stdint.h>
 #include <stdarg.h>
 #include <math.h>
 #include "Log.h"
+#include "Platform.h"
 
 #define KB 1000
 #define MB 1000000
@@ -46,7 +46,7 @@ inline bool Xor(bool a, bool b, bool c) { return (((int)a + (int)b + (int)c) == 
 inline bool Xor(bool a, bool b) { return Xor(a, b, false); }
 
 inline unsigned NumLen(int n)
-{ return n == 0 ? 1 : (unsigned) floor(log10(n) + 1); }
+{ return n == 0 ? 1 : (unsigned) floor(log10((float)n) + 1); }
 
 void BlockSignals();
